+++ HelperApp/src/main/java/com/cdw/springenablement/helperapp/exception/NoContentException.java
@@ -1,20 +1,16 @@
 package com.cdw.springenablement.helperapp.exception;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-/**
- * class which gets the message and status code and sends as a response
- */
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class HelperAppException extends RuntimeException {
+@ResponseStatus(HttpStatus.NO_CONTENT)
+public class NoContentException extends RuntimeException {
     private final HttpStatus httpStatus;
 
-    public HelperAppException(String message) {
-        this(message, HttpStatus.BAD_REQUEST);
+    public NoContentException(String message) {
+        this(message, HttpStatus.NO_CONTENT);
     }
 
-    public HelperAppException(String message, HttpStatus httpStatus) {
+    public NoContentException(String message, HttpStatus httpStatus) {
         super(message);
         this.httpStatus = httpStatus;
     }
