+++ HelperApp/src/main/java/com/cdw/springenablement/helperapp/services/interfaces/HelperService.java
@@ -1,7 +1,10 @@
 package com.cdw.springenablement.helperapp.services.interfaces;
 
 import com.cdw.springenablement.helperapp.client.models.HelperAppointmentDto;
+import com.cdw.springenablement.helperapp.client.models.TimeSlotDto;
+import com.cdw.springenablement.helperapp.client.models.TimeSlotDtos;
 
+import java.time.LocalDate;
 import java.util.List;
 /**
  * Interface which has methods of HelperService
@@ -9,4 +12,10 @@ import java.util.List;
 public interface HelperService {
 
     List<HelperAppointmentDto> getAppointment();
+
+    List<TimeSlotDto> getAvailableTechnicians(LocalDate date, Long timeSlotId) ;
+
+    List<TimeSlotDtos> getAllTimeSlots();
+
+
 }
