+++ HelperApp/src/test/java/com/cdw/springenablement/helperapp/Controller/HelperControllerTest.java
@@ -1,6 +1,8 @@
 package com.cdw.springenablement.helperapp.Controller;
 
 import com.cdw.springenablement.helperapp.client.models.HelperAppointmentDto;
+import com.cdw.springenablement.helperapp.client.models.TimeSlotDto;
+import com.cdw.springenablement.helperapp.client.models.TimeSlotDtos;
 import com.cdw.springenablement.helperapp.controller.HelperController;
 import com.cdw.springenablement.helperapp.services.interfaces.HelperService;
 import com.cdw.springenablement.helperapp.services.interfaces.UserService;
@@ -13,6 +15,8 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.web.servlet.MockMvc;
+
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -43,9 +47,28 @@ public class HelperControllerTest {
         when(helperService.getAppointment()).thenReturn(appointmentDtos);
         ResponseEntity<List<HelperAppointmentDto>> helperAppointmentDtos= helperController.getHelperAppointments();
         assertEquals(appointmentDtos,helperAppointmentDtos.getBody());
-
-
-
     }
+
+
+    @Test
+    public void testAvailableTechnicians() {
+        List<TimeSlotDto> timeSlotDtos = new ArrayList<>();
+        Long id=1L;
+        LocalDate date= LocalDate.parse("2023-09-09");
+        when(helperService.getAvailableTechnicians(date,id)).thenReturn(timeSlotDtos);
+        ResponseEntity<List<TimeSlotDto>> listResponseEntity =
+                helperController.getAvailableTechnicians(date,id);
+    }
+
+
+    @Test
+    public void  testAllTimeSlots(){
+        List<TimeSlotDtos> timeSlotDtos=new ArrayList<>();
+        when(helperService.getAllTimeSlots()).thenReturn(timeSlotDtos);
+        ResponseEntity<List<TimeSlotDtos>> timeSlotDtoss=helperController.getAllTimeSlots();
+        assertEquals(timeSlotDtos,timeSlotDtoss.getBody());
+    }
+
+
 
 }
