+++ HelperApp/src/test/java/com/cdw/springenablement/helperapp/service/UserServiceImplTest.java
@@ -85,6 +85,7 @@ public class UserServiceImplTest {
         assertEquals(1, userId);
     }
 
+
     @Test(expected = Exception.class)
     public void testRegisterUser_UserExists() throws Exception {
         UserDto userDto = new UserDto();
@@ -93,6 +94,7 @@ public class UserServiceImplTest {
         existingUser.setEmail(userDto.getEmail());
         userService.registerUser(userDto);
     }
+
 
     @Test(expected = Exception.class)
     public void testRegisterAdminUser() throws Exception {
@@ -131,15 +133,6 @@ public class UserServiceImplTest {
     }
 
 
-    @Test
-    public void testGetAvailableTechnicians_InvalidTimeslotId() {
-        LocalDate date = LocalDate.now().plusDays(1);
-        Long timeslotId = 999L;
-        HelperAppException exception = assertThrows(HelperAppException.class,
-                () -> userService.getAvailableTechnicians(date, timeslotId));
-
-        assertEquals("Invalid timeslotId provided", exception.getMessage());
-    }
 
 
     @Test
@@ -176,68 +169,6 @@ public class UserServiceImplTest {
     }
 
 
-    @Test(expected = HelperAppException.class)
-    public void testGetAvailableTechnicians_Failure() {
-        Long id = 1L;
-        List<Bookings> bookings = new ArrayList<>();
-        Bookings booking = new Bookings();
-        booking.setHelperId(id);
-        bookings.add(booking);
-        //mock data
-        LocalDate date = LocalDate.parse("2023-09-09");
-        boolean flag = false;
-        List<TimeSlot> timeSlots = new ArrayList<>();
-        TimeSlot timeSlot = new TimeSlot();
-        timeSlot.setId(id);
-        timeSlots.add(timeSlot);
-        List<TimeSlot> timeSlotDtos = new ArrayList<>();
-        if (id != null) {
-            TimeSlot timeSlot1 = new TimeSlot();
-            List<Helper> helpers = userService.getAvailableHelpersForTimeSlot(date, timeSlot, flag);
-            timeSlotDtos.add(timeSlot1);
-        }
-
-    }
-
-    @Test(expected = HelperAppException.class)
-    public void testGetAvailableTechnicians() {
-        LocalDate currentDate = LocalDate.parse("2024-09-09");
-        Long timeslotId = 1L;
-        TimeSlot mockTimeSlot = new TimeSlot();
-        mockTimeSlot.setId(timeslotId);
-        mockTimeSlot.setStartTime(LocalTime.parse("09:00"));
-        mockTimeSlot.setEndTime(LocalTime.parse("10:00"));
-        Helper mockHelper = new Helper();
-        mockHelper.setId(1L);
-        mockHelper.setSpecialization("Plumber");
-        List<Helper> availableHelpers = List.of(mockHelper);
-        when(timeSlotRepository.findById(timeslotId)).thenReturn(Optional.of(mockTimeSlot));
-        List<TimeSlotDto> result = userService.getAvailableTechnicians(currentDate, timeslotId);
-    }
-
-
-    @Test
-    public void testAvailableHelpers() {
-        List<Bookings> bookings = new ArrayList<>();
-        Long id = 1L;
-        LocalDate date = LocalDate.parse("2023-09-09");
-
-    }
-
-
-    @Test(expected = DateTimeException.class)
-    public void testGetAvailableTechniciansWithNoAvailableHelpers() {
-        List<Bookings> bookings = new ArrayList<>();
-        List<TimeSlot> timeSlots = new ArrayList<>();
-        LocalDate date = LocalDate.of(1, 2, 2002);
-        Long id = 1L;
-        userService.getAvailableTechnicians(date, id);
-    }
-
-
-
-
-
 
     @Test
     public void testGetUserBookings_Success() {
@@ -249,9 +180,7 @@ public class UserServiceImplTest {
         user.setId(id);
         user.setEmail("test@example.com");
         user.setRoles(Collections.singleton(new Roles(SuceessConstants.ROLE_RESIDENT)));
-
         when(userRepository.findByEmail(authentication.getName())).thenReturn(Optional.of(user));
-
         List<Bookings> bookings=new ArrayList<>();
         Bookings booking = new Bookings();
         booking.setId(id);
@@ -272,13 +201,8 @@ public class UserServiceImplTest {
         BookingDto bookingDtos=new BookingDto();
         bookingDtos.setBookingId(id);
         bookingDtos.setSpecialisation(bookingDtos.getSpecialisation());
-
        userService.getUserBookings();
     }
-
-
-
-
 
 
 
@@ -288,77 +212,6 @@ public class UserServiceImplTest {
         SecurityContextHolder.getContext().setAuthentication(authentication);
         userService.getUserBookings();
     }
-
-
-    @Test
-    public void testAllTimeSlots() {
-        List<TimeSlot> timeSlots = new ArrayList<>();
-        TimeSlot timeSlot = new TimeSlot();
-        long id = 1L;
-        timeSlot.setId(id);
-        timeSlot.setStartTime(LocalTime.parse("01:00"));
-        timeSlot.setEndTime(LocalTime.parse("02:00"));
-        timeSlots.add(timeSlot);
-        when(timeSlotRepository.findAll()).thenReturn(timeSlots);
-        List<TimeSlotDtos> timeSlotDtos = new ArrayList<>();
-        for (TimeSlot timeSlot1 : timeSlots) {
-            TimeSlotDtos timeSlotDtos1 = new TimeSlotDtos();
-            timeSlotDtos1.setEndTime(String.valueOf(timeSlot1.getEndTime()));
-            timeSlotDtos1.setStartTime(String.valueOf(timeSlot1.getStartTime()));
-            timeSlotDtos1.setId(id);
-            timeSlotDtos.add(timeSlotDtos1);
-        }
-        List<TimeSlotDtos> timeSlotDtos2 = userService.getAllTimeSlots();
-    }
-
-
-    @Test
-    public void testGetAvailableHelpers() {
-        List<Bookings> bookings = new ArrayList<>();
-        Long id = 1L;
-        LocalDate date = LocalDate.parse("2023-09-09");
-        Users users = new Users();
-        Boolean flag = true;
-        users.setId(id);
-        users.setEmail("pavi@gmail.com");
-        Bookings book = new Bookings();
-        book.setUsers(users);
-        book.setHelperId(id);
-        book.setDate(date);
-        List<Long> bookedHelperIds = bookings.stream()
-                .map(booking -> Long.valueOf(booking.getHelperId()))
-                .collect(Collectors.toList());
-        List<Helper> helpers = new ArrayList<>();
-        if (bookedHelperIds.isEmpty()) {
-            when(helperRepository.findAll()).thenReturn(helpers);
-        } else {
-            when(helperRepository.findByIdNotIn(bookedHelperIds)).thenReturn(helpers);
-        }
-        TimeSlot timeSlot = new TimeSlot();
-
-        List<Helper> availableHelper = userService.getAvailableHelpersForTimeSlot(date, timeSlot, flag);
-        assertEquals(availableHelper, helpers);
-    }
-
-
-    @Test
-    public void testGetAvailableTechniciansWithNullTimeslotId() {
-        LocalDate currentDate = LocalDate.now();
-        Long timeslotId = null;
-        LocalDate date = LocalDate.of(2023, 12, 12);
-        List<TimeSlot> timeSlots = new ArrayList<>();
-        timeSlots.add(new TimeSlot());
-        List<Helper> availableHelpers = new ArrayList<>();
-        availableHelpers.add(new Helper());
-        when(timeSlotRepository.findAll()).thenReturn(timeSlots);
-        List<TimeSlotDto> result = userService.getAvailableTechnicians(date, timeslotId);
-        assertNotNull(result);
-        assertEquals(1, result.size());
-    }
-
-
-
-
 
 
 }
